backend/src/requirements.txt
reboot-resemble>=0.3.2
pytest>=7.4.2
